+++ aion/web/controllers/room_controller.ex
@@ -1,7 +1,7 @@
 defmodule Aion.RoomController do
   use Aion.Web, :controller
 
-  alias Aion.{Room, ErrorCodesViews}
+  alias Aion.Room
 
   plug Guardian.Plug.EnsureAuthenticated, handler: __MODULE__
 
@@ -20,9 +20,7 @@ defmodule Aion.RoomController do
         |> put_resp_header("location", room_path(conn, :show, room))
         |> render("show.json", room: room)
       {:error, changeset} ->
-        conn
-        |> put_status(:unprocessable_entity)
-        |> render(Aion.ChangesetView, "error.json", changeset: changeset)
+        Errors.unprocessable_entity(conn, changeset)
     end
   end
 
@@ -39,9 +37,7 @@ defmodule Aion.RoomController do
       {:ok, room} ->
         render(conn, "show.json", room: room)
       {:error, changeset} ->
-        conn
-        |> put_status(:unprocessable_entity)
-        |> render(Aion.ChangesetView, "error.json", changeset: changeset)
+        Errors.unprocessable_entity(conn, changeset)
     end
   end
 
@@ -56,6 +52,6 @@ defmodule Aion.RoomController do
   end
 
   def unauthenticated(conn, _params) do
-    ErrorCodesViews.unauthenticated(conn)
+    Errors.unauthenticated(conn)
   end
 end
